Fix #27: Support cross-classpath dependencies with explicit config.
Support for cross-classpath dependencies of the form

    .dependsOn(someCrossProject)

and

    .dependsOn(someCrossProject % "test")

rely on an implicit conversion of `CrossProject` to
`CrossClasspathDependency` and `CrossClasspathDependency.Constructor`,
respectively.

However, defined in the companion object of `CrossProject`, those
implicit conversions did not have enough precedence over other
conversions in sbt, causing the `% "test"` not to compile.

This commit fixes this issue by putting the relevant implicit
conversions in `CrossPlugin.autoImport`, so that they are directly
imported in .sbt builds, therefore having the strongest precedence.