Fix #37: Move JSPlatform to org.scalajs.sbtplugin.
This is evil, as we are introducing a class in a package that is
"owned" by sbt-scalajs, but unfortunately that is the only way we
can be forward source compatible with Scala.js 1.x. Specifically,
we need this for the compatibility macro to generate code that
works with Scala.js 1.x, which will define `JSPlatform` in that
package.