Fix issues blocking Scala.js' master from migrating to sbt-crossproject (#38)
* Towards 0.2.0.

* Fix #33: Remove CrossProject.settingSets.

* Setup caching of ivy on Travis.

* Fix #27: Support cross-classpath dependencies with explicit config.

Support for cross-classpath dependencies of the form

    .dependsOn(someCrossProject)

and

    .dependsOn(someCrossProject % "test")

rely on an implicit conversion of `CrossProject` to
`CrossClasspathDependency` and `CrossClasspathDependency.Constructor`,
respectively.

However, defined in the companion object of `CrossProject`, those
implicit conversions did not have enough precedence over other
conversions in sbt, causing the `% "test"` not to compile.

This commit fixes this issue by putting the relevant implicit
conversions in `CrossPlugin.autoImport`, so that they are directly
imported in .sbt builds, therefore having the strongest precedence.

* Fix #37: Move JSPlatform to org.scalajs.sbtplugin.

This is evil, as we are introducing a class in a package that is
"owned" by sbt-scalajs, but unfortunately that is the only way we
can be forward source compatible with Scala.js 1.x. Specifically,
we need this for the compatibility macro to generate code that
works with Scala.js 1.x, which will define `JSPlatform` in that
package.