Remove println in requireDependencies
I think this will not emit the message when loading up the build.

If someone is interested in this message, the right way to implement it is in `Keys.onLoadMessage` with a way of disabling it with a special setting `crossProjectDisableLoadMessage`. This would mean creating an sbt plugin (if there's not already one).

Fix #52.