Add `crossProject.withoutSuffixFor(platform)`.
This method can be used if there is a "default" platform for the
project, which is more commonly used than the others. It will not
add the platform suffix for the given `platform`.

For example,

  val foo = crossProject(JSPlatform, JVMPlatform, NativePlatform)
    .withoutSuffixFor(JVMPlatform)
    .settings(...)

  val fooJS = foo.js
  val fooJVM = foo.jvm
  val fooNative = foo.native

will give the ID `foo` to `foo.jvm`, instead of the default `fooJVM`.
This then allows to run sbt tasks such as

  > foo/test

instead of

  > fooJVM/test

for the JVM.